Latest update for IDL compliance.  There is still a lot of optimization to do to make things faster, and make the IDL usage more streamlined. Also, the type information usually specified by a G_CanonicalPropertyType.xxx.name() doesn't have a place to be stored, and this drives the color mappings.  So colors for styling are way off.
CombinedEntitySearchPage is now pretty neutered, and you MUST provide an implementation of a SearchResultsView component in your customer implementation.  This now gives you complete control over the way things look and behave.

Added an option for turning off logging, but the CUSTOMER has to decide how/if that's honored.  i.e.  if graphene.enable-logging is false, you should use a buildLoggingDAO() static method in your DAO Module to choose the NoOp version of the logger instead of the real implementation.

Added a helper for making G_EntityQuery objects
Added a default for number formatting in case something goes wrong.
Extra null checks in many places

I have NOT tested the graph generation yet.